requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.100.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.2
python-dotenv==1.0.1
PyYAML==6.0.2
routers==0.10.1
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.27.0
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.23.2
watchfiles==1.0.4
websockets==14.1
